Can capture in any direction

+++ src/gomoku/arena/players.go
@@ -9,8 +9,8 @@ type Player interface {
 	GetId() int
 	SetColor(int8)
 	GetColor() int8
-	GetCaptured() int
-	AddCaptured(int)
+	GetCaptured() int8
+	AddCaptured(int8)
 	SetHasWon(bool)
 }
 
@@ -18,7 +18,7 @@ type DefaultPlayer struct {
 	Id       int
 	HasWon   bool
 	Color    int8
-	Captured int
+	Captured int8
 }
 
 func (dp *DefaultPlayer) SetId(newid int) {
@@ -37,11 +37,11 @@ func (dp *DefaultPlayer) GetColor() int8 {
 	return dp.Color
 }
 
-func (dp *DefaultPlayer) GetCaptured() int {
+func (dp *DefaultPlayer) GetCaptured() int8 {
 	return dp.Captured
 }
 
-func (dp *DefaultPlayer) AddCaptured(pawns int) {
+func (dp *DefaultPlayer) AddCaptured(pawns int8) {
 	dp.Captured += pawns
 }
